tests: robustness enhancement of regression tests
Results of some regression tests may vary on diffirent env,
due to the unready status of shd process or unexpected
output language of key commands in test case.

So, it should be made sure that shd process has got ready before
a `gluster v heal` cmd is triggered in testcase. And if the output
of command is the judge standard of sub-testcase, we also shuold
assure it is shown in specified language as expected.

Fixes: #4432
Signed-off-by: chenjinhao <chen.jinhao@zte.com.cn>